fix(Backend): 三轮扫描修复—订单状态流转、null安全、SQL优化、校验增强
- Critical: nextStatusOrder增加状态流转校验，防止跳过洗衣流程
- Critical: calculationOrder支持userCouponId=0不使用优惠券
- Critical: updateOrderStatus统一为所有READY_FOR_PICKUP路径分配寄存柜
- Critical: shippingOrder目标状态从WASHING修正为RECEIVED
- High: UpdateUserFrom移除balance字段防止余额篡改
- High: AddCouponFrom增加discount/threshold正数校验
- High: UpdateAdminUserFrom增加username非空校验
- High: RechargeRecordsServiceImpl增加用户存在性校验
- High: DashboardMapper/Schools/Lockers服务层统一增加null检查
- Medium: ExceptionControllerAdvice业务异常降级为warn
- Medium: LockersMapper.unLocker增加last_used_at更新
- Medium: UserCouponMapper.xml修复threshold表别名和status字符串比较
- Low: OrdersController PostMapping补全前导斜杠
- DB: V1/V3迁移修复payments.paid_at和roles.created_at时间戳

+++ SmartWash/src/main/java/com/smartwash/from/order/ReservationOrderFrom.java
@@ -3,9 +3,11 @@ package com.smartwash.from.order;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReservationOrderFrom {
     @NotNull(message = "请选择套餐")
     private Long itemsId;
-    private Float totalPrice;
+    private BigDecimal totalPrice;
 }
